Manage subprocess output in real-time.
Move subprocess management into crepuscular.proclib and use a select loop,
allowing us to interact with child process output in real-time.  This lets us
see the output of tools like gcloud in real-time.  Also make it so that a
hook can output multiple json objects, one per line on standard output, to be
sent back to the framework.